fix(masthead): move search actions out of combobox
